moved addition column creation to imputation
