+++ src/imputation/imputation_helpers.py
@@ -1,5 +1,6 @@
 """Utility functions  to be used in the imputation module."""
 import logging
+import os
 import pandas as pd
 from typing import List, Tuple
 from itertools import chain
@@ -440,3 +441,75 @@ def create_new_backdata(backdata: pd.DataFrame, config) -> pd.DataFrame:
     wanted_cols = list(schema.keys())
 
     return backdata[wanted_cols]
+
+
+def identify_key_business(df):
+    """ Function to identify the key business using reference column & key
+    busineses lookup table.
+    Args:
+        df (pd.DataFrame): The dataframe to identify the key business columns.
+    Return:
+        df (pd.DataFrame): The dataframe with the identified key business columns.
+    """
+    # get key businesses
+    rootpath = "R:/BERD Results System Development 2023/DAP_emulation/2021_surveys/PNP/"
+
+    key_businesses_df = pd.read_csv(os.path.join(rootpath, 'KEYS 2023.csv'))
+    key_businesses_list = list(key_businesses_df["2023 KEYS"])
+
+    df['pnp_key'] = df['reference'].apply(
+        lambda x: 'Key0' if x in key_businesses_list else 'Key1'
+    )
+
+    return df
+
+
+def identify_osmotherly_businesses(df):
+    """ Function to identify the osmotherly businesses using reference
+    column & osmotherly busineses lookup table.
+    Args:
+        df (pd.DataFrame): The dataframe to identify the osmotherly business
+        columns.
+    Return:
+        df (pd.DataFrame): The dataframe with the identified osmotherly business
+        columns.
+    """
+    # get osmotherly businesses
+    rootpath = "R:/BERD Results System Development 2023/DAP_emulation/2021_surveys/PNP/"
+
+    osmotherly_businesses_df = pd.read_csv(os.path.join(rootpath,
+                                                        "Osmotherly PNP 2023.csv"))
+    osmotherly_businesses_list = list(osmotherly_businesses_df["ruref"])
+
+    df['osmotherly'] = df['reference'].apply(
+        lambda x: True if x in osmotherly_businesses_list else False
+    )
+
+    return df
+
+
+def add_area_column(df):
+    """
+    Add area columns to the dataframe bassed on ITL121NM column.
+
+    Args:
+        df (pd.DataFrame): The dataframe to add the area columns to.
+
+    Returns:
+        pd.DataFrame: The dataframe with the area columns added.
+    """
+    area_dict = {'South East (England)': 'select',
+                 'East': 'select',
+                 'London': 'select',
+                 'West Midlands (England)': 'other',
+                 'Wales': 'other',
+                 'North East (England)': 'other',
+                 'East Midlands (England)': 'other',
+                 'Yorkshire and The Humber': 'other',
+                 'Scotland': 'other',
+                 'South West (England)': 'other',
+                 'North West (England)': 'other'}
+
+    df['area'] = df['ITL121NM'].map(area_dict)
+
+    return df
